+++ packages/parser/src/index.ts
@@ -1,3 +1,4 @@
 export { version as VTJ_PARSER_VERSION } from './version';
 export * from './uni';
 export * from './vue';
+export * from './tools';