+++ packages/parser/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.56](https://gitee.com/newgateway/vtj/compare/@vtj/parser@0.12.55...@vtj/parser@0.12.56) (2025-07-16)
+
+
+### Bug Fixes
+
+* 🐛 parser 数据源 ([9befb29](https://gitee.com/newgateway/vtj/commits/9befb2968ecc74f33b01202317e966c8f3bbadcd))
+* 🐛 parser watch ([9404146](https://gitee.com/newgateway/vtj/commits/940414637c3c079014f3d1c0f5dfcd53edf2f693))
+
+
+### Features
+
+* ✨ 代码校验并自动修复 ([205257b](https://gitee.com/newgateway/vtj/commits/205257b75b5ebcb5c02ec4ade0000388552277a6))
+* ✨ parser validator and fixer ([0a7e29c](https://gitee.com/newgateway/vtj/commits/0a7e29cee2d89fce800521c43af989252d937ea3))
+* ✨ parser watch ([26a942a](https://gitee.com/newgateway/vtj/commits/26a942a1b1297fa7d5c7ad49121b9175efce5b62))
+
+
+
+
+
+## [0.12.55](https://gitee.com/newgateway/vtj/compare/@vtj/parser@0.12.54...@vtj/parser@0.12.55) (2025-07-09)
+
+**Note:** Version bump only for package @vtj/parser
+
+
+
+
+
+## [0.12.54](https://gitee.com/newgateway/vtj/compare/@vtj/parser@0.12.53...@vtj/parser@0.12.54) (2025-07-08)
+
+
+### Bug Fixes
+
+* 🐛 升级依赖 ([dbee35b](https://gitee.com/newgateway/vtj/commits/dbee35bd867a44f8c71c117fa90d56d108144a6b))
+
+
+
+
+
 ## [0.12.53](https://gitee.com/newgateway/vtj/compare/@vtj/parser@0.12.52...@vtj/parser@0.12.53) (2025-07-07)
 
 **Note:** Version bump only for package @vtj/parser