+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.56](https://gitee.com/newgateway/vtj/compare/@vtj/utils@0.12.55...@vtj/utils@0.12.56) (2025-07-16)
+
+**Note:** Version bump only for package @vtj/utils
+
+
+
+
+
+## [0.12.55](https://gitee.com/newgateway/vtj/compare/@vtj/utils@0.12.54...@vtj/utils@0.12.55) (2025-07-09)
+
+**Note:** Version bump only for package @vtj/utils
+
+
+
+
+
+## [0.12.54](https://gitee.com/newgateway/vtj/compare/@vtj/utils@0.12.53...@vtj/utils@0.12.54) (2025-07-08)
+
+
+### Bug Fixes
+
+* 🐛 升级依赖 ([dbee35b](https://gitee.com/newgateway/vtj/commits/dbee35bd867a44f8c71c117fa90d56d108144a6b))
+
+
+
+
+
 ## [0.12.53](https://gitee.com/newgateway/vtj/compare/@vtj/utils@0.12.52...@vtj/utils@0.12.53) (2025-07-07)
 
 **Note:** Version bump only for package @vtj/utils